Spartans/requirements.txt
torch
opencv-python
scikit-learn
matplotlib
pandas
numpy
ipywidgets
Pillow
